refactor: GamePage에서 사용되는 컴포넌트를 페이지 디렉토리 내부로 이동 #276
