+++ frontend/src/components/layout/Header/Header.tsx
@@ -22,11 +22,11 @@ import SettingIcon from '@/assets/images/settingIcon.svg';
 import A11yOnly from '@/components/common/a11yOnly/A11yOnly';
 import AlertModal from '@/components/common/AlertModal/AlertModal';
 import RoomSettingModal from '@/components/common/RoomSettingModal/RoomSettingModal';
-import { convertMsecToSecond } from '@/components/SelectContainer/Timer/Timer.util';
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
 import useFocus from '@/hooks/useFocus';
 import useGetUserInfo from '@/hooks/useGetUserInfo';
 import useModal from '@/hooks/useModal';
+import { convertMsecToSecond } from '@/pages/GamePage/SelectContainer/Timer/Timer.util';
 
 interface HeaderProps {
   title: string;
