+++ frontend/src/pages/GamePage/GamePage.tsx
@@ -1,6 +1,7 @@
+import SelectContainer from './SelectContainer/SelectContainer';
+
 import Content from '@/components/layout/Content/Content';
 import { GameHeader } from '@/components/layout/Header/Header';
-import SelectContainer from '@/components/SelectContainer/SelectContainer';
 import TopicContainer from '@/components/TopicContainer/TopicContainer';
 
 const GamePage = () => {
