+++ frontend/src/pages/GamePage/SelectContainer/SelectContainer.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 import useSelectOption from './hooks/useSelectOption';
 import { selectContainerLayout, selectSection } from './SelectContainer.styled';
 import Timer from './Timer/Timer';
-import SelectButton from '../common/SelectButton/SelectButton';
+import SelectButton from '../SelectButton/SelectButton';
 
 import SelectOption from '@/components/SelectOption/SelectOption';
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
