+++ frontend/src/pages/GamePage/SelectContainer/Timer/Timer.styled.ts
