+++ frontend/src/pages/GamePage/SelectContainer/hooks/useVoteIsFinished.ts
