+++ frontend/src/pages/GamePage/SelectContainer/hooks/useVoteIsFinishedQuery.ts
